Problems related with the field "workratio"

+++ src/main/java/com/lesstif/jira/issue/IssueFields.java
@@ -70,7 +70,7 @@ public class IssueFields {
 	
 	private String[] labels;
 	
-	private Integer workratio;
+	private Long workratio;
 	
 	private String environment;
